Added websocket client to platform services.
